requirements.txt
fastapi
uvicorn[standard]
sqlmodel
yfinance
python-binance
google-genai
requests
python-dotenv
pydantic-settings
pydantic>=2.0.0
python-multipart
pyotp
python-jose[cryptography]
passlib[bcrypt]
